chore: Wip

+++ nflux-ebpf/src/maps.rs
@@ -19,5 +19,5 @@ pub static TC_CONFIG: Array<Configmap> = Array::with_max_entries(1, 0);
 pub static TC_EVENT: RingBuf = RingBuf::with_byte_size(4096, 0);
 
 #[map]
-pub static ACTIVE_CONNECTIONS: LruHashMap<ActiveConnectionKey, u8> =
+pub static ACTIVE_CONNECTIONS: LruHashMap<ActiveConnectionKey, u64> =
     LruHashMap::with_max_entries(4096, 0);
